Removed prints

+++ simple_train.py
@@ -78,4 +78,5 @@ def main(config):
     train_dataset = load_dataset(config.dataset_id, split=config.split)
 
     model = AutoModelForCausalLM.from_pretrained(
-        config.model_id
+        config.model_id,
+        torch_dtype=torch.bfloat16,
